+++ src/main/java/matteroverdrive/blocks/BlockMatterAnalyzer.java
@@ -18,17 +18,11 @@
 
 package matteroverdrive.blocks;
 
-import matteroverdrive.MatterOverdrive;
 import matteroverdrive.blocks.includes.MOBlockMachine;
 import matteroverdrive.machines.analyzer.TileEntityMachineMatterAnalyzer;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.inventory.InventoryHelper;
-import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 import javax.annotation.Nonnull;
@@ -115,51 +109,4 @@ public class BlockMatterAnalyzer extends MOBlockMachine<TileEntityMachineMatterA
     {
         return MOBlockRenderer.renderID;
     }*/
-
-    public static void setState(boolean active, World worldIn, BlockPos pos) {
-        IBlockState iBlockState = worldIn.getBlockState(pos);
-
-        TileEntity tileentity = worldIn.getTileEntity(pos);
-
-        keepInventory = true;
-
-        if (active) {
-            worldIn.setBlockState(pos, MatterOverdrive.BLOCKS.matter_analyzer_on.getDefaultState().withProperty(PROPERTY_DIRECTION, iBlockState.getValue(PROPERTY_DIRECTION)), 3);
-        } else {
-            worldIn.setBlockState(pos, MatterOverdrive.BLOCKS.matter_analyzer_off.getDefaultState().withProperty(PROPERTY_DIRECTION, iBlockState.getValue(PROPERTY_DIRECTION)), 3);
-        }
-
-        keepInventory = false;
-
-        if (tileentity != null) {
-            tileentity.validate();
-
-            worldIn.setTileEntity(pos, tileentity);
-        }
-    }
-
-    @Override
-    public IBlockState getStateForPlacement(World world, BlockPos pos, EnumFacing facing, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer)
-    {
-        return this.getDefaultState().withProperty(PROPERTY_DIRECTION, placer.getHorizontalFacing().getOpposite());
-    }
-
-    @Override
-    public ItemStack getItem(World worldIn, BlockPos pos, IBlockState state) {
-        return new ItemStack(MatterOverdrive.BLOCKS.matter_analyzer_off);
-    }
-
-    @Override
-    public void breakBlock(World worldIn, BlockPos pos, IBlockState state)
-    {
-        if (!keepInventory) {
-            TileEntityMachineMatterAnalyzer tileentity = (TileEntityMachineMatterAnalyzer) worldIn.getTileEntity(pos);
-
-            if (tileentity != null) {
-                InventoryHelper.dropInventoryItems(worldIn, pos, tileentity);
-                worldIn.updateComparatorOutputLevel(pos, this);
-                super.breakBlock(worldIn, pos, state);
-            }
-        }
-    }
 }